Extract `QrPainter` and reuse it in automotive and tv modules
