+++ automotive/build.gradle.kts
@@ -96,7 +96,6 @@ dependencies {
     implementation(libs.rx2.java)
     implementation(libs.timber)
     implementation(libs.work.runtime)
-    implementation(libs.zxing)
 
     implementation(projects.modules.features.account)
     implementation(projects.modules.features.cartheme)
