+++ modules/services/compose/build.gradle.kts
@@ -41,6 +41,7 @@ dependencies {
     implementation(libs.lottie.compose)
     implementation(libs.navigation.compose)
     implementation(libs.reorderable)
+    implementation(libs.zxing)
 
     implementation(projects.modules.services.images)
     implementation(projects.modules.services.localization)
